requirements.txt
Flask
python-dotenv
firebase-admin
requests
boto3
flask-cors
supabase-py
